Poi Simulator with programmable lights added
Two poi toys for the hands of each stickman. There's a little pattern
function array and a factory for text rendering. You can set the number
of lights for each kinetic rope. The rope simulation is approximated
with three connected springs and a Bernstein-polynomial of 6th degree.
The light streaks are rendered using canvas 2d lines. There are two
separated canvas 2d scenes available as textures in the shader programs
now. sampler_poitoy_scene for the light streaks and sampler_sticky_scene
for the stickmen. Additive blending in the poi toy texture in the
advance step with a fade-out makes a persistence of vision effect.